requirements.txt
# IdentityCrisis Bot + Dashboard Dependencies

# Discord API wrapper
discord.py>=2.3.0

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Templates
jinja2>=3.1.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0

# HTTP client (for Discord OAuth)
httpx>=0.26.0

# Environment variables
python-dotenv>=1.0.0
